Conflicts:
	MyApplication2/app/build.gradle
	MyApplication2/app/src/main/java/com/example/dllo/mirror/controller/fragment/PageFragment.java

+++ MyApplication2/app/src/main/java/com/example/dllo/mirror/controller/activity/SignUpActivity.java
@@ -13,6 +13,7 @@ import com.example.dllo.mirror.model.bean.UserBean;
 import com.example.dllo.mirror.model.db.Users;
 import com.example.dllo.mirror.model.db.UsersDao;
 import com.example.dllo.mirror.model.utils.GreenDaoSingleton;
+import com.example.dllo.mirror.model.utils.OkHttpClientManager;
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import com.squareup.okhttp.Callback;
@@ -24,6 +25,7 @@ import com.squareup.okhttp.Response;
 
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.Map;
 
 import cn.sharesdk.framework.Platform;
 import cn.sharesdk.framework.PlatformActionListener;
@@ -44,6 +46,8 @@ public class SignUpActivity extends BaseActivity implements View.OnClickListener
     private Platform platform;
     private UsersDao usersDao;
     ImageView closeIv;
+    private Gson gson;
+    OkHttpClientManager clientManager;
 
 
     @Override
@@ -71,6 +75,8 @@ public class SignUpActivity extends BaseActivity implements View.OnClickListener
 
     @Override
     protected void initData() {
+
+        clientManager = OkHttpClientManager.getInstance();
         closeIv.setOnClickListener(this);
 
         creatUserButton.setOnClickListener(this);
@@ -95,37 +101,56 @@ public class SignUpActivity extends BaseActivity implements View.OnClickListener
 
                 String postUrl1 = "http://api101.test.mirroreye.cn/index.php/user/login";
 
-                //请求体
-                FormEncodingBuilder builder1 = new FormEncodingBuilder();
-                builder1.add("phone_number", phoneEt.getText().toString());
-                builder1.add("password", passwordEt.getText().toString());
-
-
-                //创建请求体
-                RequestBody body1 = builder1.build();
-                //创建请求
-                Request request1 = new Request.Builder().url(postUrl1).post(body1).build();
-
-                //发出异步post请求
-                client.newCall(request1).enqueue(new Callback() {
+                Map<String, String> map = new HashMap<>();
+                map.put("phone_number", phoneEt.getText().toString());
+                map.put("password", passwordEt.getText().toString());
+                clientManager.postAsyn(postUrl1, new OkHttpClientManager.ResultCallback<UserBean>() {
                     @Override
-                    public void onFailure(Request request, IOException e) {
-
-                        Log.d("SignUpActivity", "error");
+                    public void onError(Request request, Exception e) {
 
                     }
 
                     @Override
-                    public void onResponse(Response response) throws IOException {
+                    public void onResponse(UserBean response) {
 
-                        Log.d("SignUpActivity", response.body().string());
+                        Log.d("SignUpActivity", "response.getData():" + response.getResult());
 
-
-                        addDb(response.body().string());
-
+                        addDb(response);
+                        finish();
 
                     }
-                });
+                }, map);
+
+
+//                //请求体
+//                FormEncodingBuilder builder1 = new FormEncodingBuilder();
+//                builder1.add("phone_number", phoneEt.getText().toString());
+//                builder1.add("password", passwordEt.getText().toString());
+//
+//
+//                //创建请求体
+//                RequestBody body1 = builder1.build();
+//                //创建请求
+//                Request request1 = new Request.Builder().url(postUrl1).post(body1).build();
+//
+//                //发出异步post请求
+//                client.newCall(request1).enqueue(new Callback() {
+//                    @Override
+//                    public void onFailure(Request request, IOException e) {
+//
+//                        Log.d("SignUpActivity", "error");
+//
+//                    }
+//
+//                    @Override
+//                    public void onResponse(Response response) throws IOException {
+//
+//                        Log.d("SignUpActivity", response.body().string());
+//
+//                        addDb(response.body().string());
+//
+//                    }
+//                });
 
 
                 break;
@@ -160,6 +185,8 @@ public class SignUpActivity extends BaseActivity implements View.OnClickListener
         platform.setPlatformActionListener(new PlatformActionListener() {
             @Override
             public void onComplete(Platform platform, int i, HashMap<String, Object> hashMap) {
+
+
                 Log.d("SignUpActivity", "success");
                 Log.d("SignUpActivity", platform.getDb().getUserId());
                 Log.d("SignUpActivity", platform.getDb().getToken());
@@ -187,12 +214,19 @@ public class SignUpActivity extends BaseActivity implements View.OnClickListener
 
     }
 
-    private void addDb(String response) {
-        Gson gson = new Gson();
-        UserBean bean = gson.fromJson(response, new TypeToken<UserBean>() {
-        }.getType());
-        usersDao.deleteAll();
-        usersDao.insert(new Users(1l, bean.getData().getUid(), bean.getData().getToken()));
+    private void addDb(UserBean response) {
+
+
+        if (!response.equals("")) {
+
+            gson = new Gson();
+
+            usersDao.deleteAll();
+            usersDao.insert(new Users(1l, response.getData().getUid(), response.getData().getToken()));
+
+            Log.d("SignUpActivity", response.getData().getUid());
+            Log.d("SignUpActivity", response.getData().getToken());
+        }
     }
